humility-core: core: breakout each core to its own file.
